// src/app/modules/master/models/lov.ts
export class LOV {
    id: string;
    created_dt: Date;
    created_by: string;
    modified_dt: Date;
    modified_by: string;
    is_deleted: boolean;
    name: string;
    value: string;
    description: string;
    par_id: string;
    lov_type_id: string;
    strvalue_1: string;
    strvalue_2: string;
    strvalue_3: string;
    numvalue_1: number;
    numvalue_2: number;
    numvalue_3: number;
}